Laborator cu tema( Am avut cateva probleme cu "ValidationBehaviour" si a trebuit sa il scot

+++ ProductPage.xaml.cs
@@ -1,5 +1,5 @@
 using CataramaAlexandruLab7.Models;
-using CataramaAlexandruLab7.Data;
+
 namespace CataramaAlexandruLab7;
 
 public partial class ProductPage : ContentPage
@@ -11,38 +11,58 @@ public partial class ProductPage : ContentPage
         sl = slist;
     }
 
+
     async void OnSaveButtonClicked(object sender, EventArgs e)
     {
         var product = (Product)BindingContext;
         await App.Database.SaveProductAsync(product);
         listView.ItemsSource = await App.Database.GetProductsAsync();
     }
-
     async void OnDeleteButtonClicked(object sender, EventArgs e)
     {
         var product = listView.SelectedItem as Product;
-        await App.Database.DeleteProductAsync(product);
-        listView.ItemsSource = await App.Database.GetProductsAsync();
+
+        if (product != null)
+        {
+            await App.Database.DeleteProductAsync(product);
+            listView.ItemsSource = await App.Database.GetProductsAsync();
+        }
+        else
+        {
+            await DisplayAlert("Error", "Please select a product to delete.", "OK");
+        }
     }
     protected override async void OnAppearing()
     {
         base.OnAppearing();
         listView.ItemsSource = await App.Database.GetProductsAsync();
     }
+
+
     async void OnAddButtonClicked(object sender, EventArgs e)
     {
-        Product p;
         if (listView.SelectedItem != null)
         {
-            p = listView.SelectedItem as Product;
-            var lp = new ListProduct()
+            Product p = listView.SelectedItem as Product;
+
+            if (p != null)
             {
-                ShopListID = sl.ID,
-                ProductID = p.ID
-            };
-            await App.Database.SaveListProductAsync(lp);
-            p.ListProducts = new List<ListProduct> { lp };
-            await Navigation.PopAsync();
+                var lp = new ListProduct
+                {
+                    ShopListID = sl.ID,
+                    ProductID = p.ID
+                };
+
+                await App.Database.SaveListProductAsync(lp);
+                p.ListProducts = new List<ListProduct> { lp };
+
+                await Navigation.PopAsync();
+            }
+        }
+        else
+        {
+            await DisplayAlert("Error", "Please select a product to add.", "OK");
         }
     }
+
 }
